+++ test-proj-configs
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from tpc_utils import check_minimum_python_version, GoodFormatter
+from utils import check_minimum_python_version, GoodFormatter
 check_minimum_python_version(3, 4)
 
 import argparse, sys, pathlib
